added weired algorithm from CSES

+++ algorithmic_parts/weirdAlgorithm.cpp
@@ -2,20 +2,20 @@
 
 using namespace std;
 
-void weiredAlgorithm(int);
+void weiredAlgorithm(unsigned long);
 
 int main(){
-    int n;
+    unsigned long n;
     cin >> n;
   weiredAlgorithm(n);
   return 0;
 }
 
-void weiredAlgorithm(int n){
+void weiredAlgorithm(unsigned long n){
   cout << n << " ";
   if(n == 1) return;
   else if(n % 2 == 0){
-    int t = n / 2;
+    unsigned long t = n / 2;
     return weiredAlgorithm(t);
   }
   else weiredAlgorithm(n*3+1);
